devops-projects/techshop-v1/backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-CORS